Reuse & export the helper that generates the config for device state GET
Change-type: minor
